Attempt more usable projectile find file config
It's really pretty slow on our repo without caching, however, the
cache goes sideways when branches change and it's annoying to have to
remember to refresh it manually. In theory this magit advice should
rebuild the cache sensibly most of the time, although I am not sure it
catches all relevant git actions so there may be some manual
rebuilding still.